+++ app/src/main/java/eu/kanade/presentation/util/Navigator.kt
@@ -68,6 +68,7 @@ interface AssistContentScreen {
 fun DefaultNavigatorScreenTransition(
     navigator: Navigator,
     modifier: Modifier = Modifier,
+    content: ScreenTransitionContent = { it.Content() },
 ) {
     val slideDistance = rememberSlideDistance()
     ScreenTransition(
@@ -79,6 +80,7 @@ fun DefaultNavigatorScreenTransition(
             )
         },
         modifier = modifier,
+        content = content,
     )
 }